+++ src/com/gtdclan/signtimer/Times.java
@@ -27,19 +27,20 @@ public class Times {
 		else {
 			Long savedTime = data.getTime();
 			
+			String st_time = this.plugin.Util.formatTime(Integer.valueOf(time));
 			if (time < savedTime) {
-				message = "New personal best. " + this.plugin.Util.getFriendly(Integer.valueOf(time));
-				bmessage = "with a new personal best of " + this.plugin.Util.getFriendly(Integer.valueOf(time));
+				message = "New personal best. " + st_time;
+				bmessage = "with a new personal best of " + st_time;
 				data.setTime(time);
 				this.plugin.database.getDatabase().save(data);
 			}
 			else if (time == savedTime) {
-				message = "You matched your best time of " + this.plugin.Util.getFriendly(Integer.valueOf(time));
-				bmessage = "They matched their best time of " + this.plugin.Util.getFriendly(Integer.valueOf(time));
+				message = "You matched your best time of " + st_time;
+				bmessage = "They matched their best time of " + st_time;
 			}
 			else {
 				Integer difference = (int) (time - savedTime);
-				String fasterTime = this.plugin.Util.getFriendly(difference);
+				String fasterTime = this.plugin.Util.formatTime(difference);
 				message = "Your old time was " + fasterTime + " faster.";
 				bmessage = "which was " + fasterTime + " slower.";
 			}
@@ -55,12 +56,27 @@ public class Times {
 		
 	}
 	
+	public void clearTimes() {
+		Query<DB> data = this.plugin.database.getDatabase().find(DB.class).where().gt("id", 0).orderBy("id");
+		int timeCount = data.findRowCount();
+		if (data != null && timeCount > 0) {
+			List<DB> times = data.findList();
+			for (DB time : times) {
+				this.plugin.database.getDatabase().delete(time);
+			}
+		}
+		else {
+			this.plugin.log.log(Level.INFO, "Error: could not find any times in DB to remove.");
+		}
+		this.plugin.rankList.clear();
+	}
+	
 	public void getTime(String playerName) {
 		String message;
 		String playerInfo = this.plugin.rankList.get(playerName);
 		if (playerInfo != null) {
-			String[] pInfo = this.plugin.rankList.get(playerName).split(",");
-			String playerTime = this.plugin.Util.getFriendly(Integer.valueOf(pInfo[1]));
+			String[] pInfo = playerInfo.split(",");
+			String playerTime = this.plugin.Util.formatTime(Integer.valueOf(pInfo[1]));
 			
 			message = " " + String.format("At rank # %02d", Integer.valueOf(pInfo[0])) + ", your best time is " + playerTime;
 		}
@@ -79,14 +95,19 @@ public class Times {
 		if (this.plugin.rankList.size() > 0) {
 			if (top) {
 				
-				this.plugin.Util.messagePlayer(senderName, "^gold^underline  Rank  |  Name  |  Time");
+				// this.plugin.Util.messagePlayer(senderName, "^gold^underline  #                     Time                        Name         ");
+				this.plugin.Util.messagePlayer(senderName, "Top 10 Times:");
 				
 				int i = 0;
 				for (String playerName : this.plugin.rankList.keySet()) {
 					String[] playerInfo = this.plugin.rankList.get(playerName).split(",");
-					String playerTime = this.plugin.Util.getFriendly(Integer.valueOf(playerInfo[1]));
-					
-					this.plugin.Util.messagePlayer(senderName, "  " + String.format("%02d", Integer.valueOf(playerInfo[0])) + "  |  " + playerName + "  |  " + playerTime);
+					String playerTime = this.plugin.Util.formatTime(Integer.valueOf(playerInfo[1]));
+					if (this.plugin.compact_time) {
+						this.plugin.Util.messagePlayer(senderName, "  " + String.format("%02d", Integer.valueOf(playerInfo[0])) + "  |  " + playerTime + "  |  " + playerName);
+					}
+					else {
+						this.plugin.Util.messagePlayer(senderName, "  " + String.format("%02d", Integer.valueOf(playerInfo[0])) + "  |  " + playerName + "  |  " + playerTime);
+					}
 					i++;
 					if (i > 9) {
 						break;
@@ -96,8 +117,8 @@ public class Times {
 			else {
 				String playerInfo = this.plugin.rankList.get(rankPlayer);
 				if (playerInfo != null) {
-					String[] pInfo = this.plugin.rankList.get(rankPlayer).split(",");
-					String playerTime = this.plugin.Util.getFriendly(Integer.valueOf(pInfo[1]));
+					String[] pInfo = playerInfo.split(",");
+					String playerTime = this.plugin.Util.formatTime(Integer.valueOf(pInfo[1]));
 					
 					this.plugin.Util.messagePlayer(senderName, " " + String.format(rankPlayer + " is rank # %02d", Integer.valueOf(pInfo[0])) + ", with a time of " + playerTime);
 				}
@@ -113,6 +134,7 @@ public class Times {
 	
 	public void updateRanks() {
 		Query<DB> data = this.plugin.database.getDatabase().find(DB.class).where().gt("id", 0).order().asc("Time");
+		
 		int timeCount = data.findRowCount();
 		if (data != null && timeCount > 0) {
 			List<DB> timeList = data.findList();