+++ CHANGELOG.md
@@ -1,2 +1,9 @@
+# v1.1.0 - Works with CB 1.3.2-R0.2 - CB 1.4.6-R0.2+
+- Added: command and permission to clear all times. defaults to OP.
+- Added: permission to allow players to remove sign timers instead of just OPs. defaults to OP.
+- Added: permission to allow players to create sign timers. defaults to OP.
+- Added: Compact time format and config setting to disable it.
+- Added: MCStats integrations. See Server Owners Section at https://mcstats.org/learn-more/ to opt-out
+ 
 # v1.0.0
 - Initial release.